pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "nebulus-atom"
version = "0.1.0"
description = "A professional, autonomous AI engineer CLI."
readme = "README.md"
authors = [
  { name = "J.L. West", email = "jlwestsr@example.com" },
]
requires-python = ">=3.12"
dependencies = [
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "typer>=0.9.0",
    "rich>=13.0.0",
    "discord.py>=2.3.0",
    "textual>=0.47.0",
    "prompt_toolkit>=3.0.0",
    "streamlit>=1.30.0",
    "pandas>=2.0.0",
]

[project.scripts]
nebulus-atom = "nebulus_atom.main:app"
mn = "nebulus_atom.main:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["nebulus_atom*"]
